+++ datumbox-framework-core/src/main/java/com/datumbox/framework/core/common/text/tokenizers/AbstractTokenizer.java
@@ -1,0 +1,35 @@
+/**
+ * Copyright (C) 2013-2017 Vasilis Vryniotis <bbriniotis@datumbox.com>
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.datumbox.framework.core.common.text.tokenizers;
+
+import java.util.List;
+
+/**
+ * This abstract class should be implemented by classes that are responsible to get
+ * a string as input and separate it into tokens/keywords.
+ * 
+ * @author Vasilis Vryniotis <bbriniotis@datumbox.com>
+ */
+public abstract class AbstractTokenizer {
+    
+    /**
+     * The tokenize method accepts a string as input and returns a list of tokens.
+     * 
+     * @param text
+     * @return 
+     */
+    public abstract List<String> tokenize(String text);
+}