+++ datumbox-framework-core/src/main/java/com/datumbox/framework/core/machinelearning/MLBuilder.java
@@ -17,7 +17,7 @@ package com.datumbox.framework.core.machinelearning;
 
 import com.datumbox.framework.common.Configuration;
 import com.datumbox.framework.common.interfaces.Parameterizable;
-import com.datumbox.framework.common.interfaces.Trainable;
+import com.datumbox.framework.core.machinelearning.common.interfaces.Trainable;
 
 import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;