+++ datumbox-framework-core/src/main/java/com/datumbox/framework/core/machinelearning/common/abstracts/AbstractTrainer.java
@@ -17,7 +17,7 @@ package com.datumbox.framework.core.machinelearning.common.abstracts;
 
 import com.datumbox.framework.common.Configuration;
 import com.datumbox.framework.common.dataobjects.Dataframe;
-import com.datumbox.framework.common.interfaces.Trainable;
+import com.datumbox.framework.core.machinelearning.common.interfaces.Trainable;
 import com.datumbox.framework.common.storageengines.abstracts.BigMapHolder;
 import com.datumbox.framework.common.storageengines.interfaces.StorageEngine;
 import com.datumbox.framework.common.utilities.RandomGenerator;