+++ datumbox-framework-core/src/main/java/com/datumbox/framework/core/machinelearning/common/dataobjects/TrainableBundle.java
@@ -16,7 +16,7 @@
 package com.datumbox.framework.core.machinelearning.common.dataobjects;
 
 import com.datumbox.framework.common.interfaces.Savable;
-import com.datumbox.framework.common.interfaces.Trainable;
+import com.datumbox.framework.core.machinelearning.common.interfaces.Trainable;
 import com.datumbox.framework.core.machinelearning.common.interfaces.Parallelizable;
 
 import java.util.HashMap;